requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
snowflake-connector-python==3.5.0
python-dotenv==1.0.0
openai==1.3.0
python-jose[cryptography]==3.3.0
pydantic==2.4.2